Organized and added visual studio proj
