+++ inc/token_info.h
@@ -1,0 +1,49 @@
+/*
+Software License Agreement (BSD License)
+
+Copyright (c) 2006, Luke Jennings (0xlukej@gmail.com)
+All rights reserved.
+
+Redistribution and use of this software in source and binary forms, with or without modification, are
+permitted provided that the following conditions are met:
+
+* Redistributions of source code must retain the above
+  copyright notice, this list of conditions and the
+  following disclaimer.
+
+* Redistributions in binary form must reproduce the above
+  copyright notice, this list of conditions and the
+  following disclaimer in the documentation and/or other
+  materials provided with the distribution.
+
+* Neither the name of Luke Jennings nor the names of its
+  contributors may be used to endorse or promote products
+  derived from this software without specific prior
+  written permission of Luke Jennings.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED
+WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR
+TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
+ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+*/
+#ifndef INC_TOKEN_INFO_H
+#define INC_TOKEN_INFO_H
+
+BOOL is_delegation_token(HANDLE token);
+BOOL is_impersonation_token(HANDLE token);
+BOOL is_token(HANDLE token, char *requested_name);
+BOOL is_local_system();
+BOOL has_impersonate_priv(HANDLE hToken);
+BOOL has_assignprimarytoken_priv(HANDLE hToken);
+DWORD TryEnableDebugPriv(HANDLE hToken);
+DWORD TryEnableAssignPrimaryPriv(HANDLE hToken);
+
+BOOL get_domain_username_from_token(HANDLE token, char *full_name_to_return);
+BOOL get_domain_groups_from_token(HANDLE token, char **group_name_array[], DWORD *num_groups);
+BOOL get_domain_from_token(HANDLE token, char *domain_to_return);
+
+#endif
