--- list_tokens.h
@@ -1,75 +1,0 @@
-/*
-Software License Agreement (BSD License)
-
-Copyright (c) 2006, Luke Jennings (0xlukej@gmail.com)
-All rights reserved.
-
-Redistribution and use of this software in source and binary forms, with or without modification, are
-permitted provided that the following conditions are met:
-
-* Redistributions of source code must retain the above
-  copyright notice, this list of conditions and the
-  following disclaimer.
-
-* Redistributions in binary form must reproduce the above
-  copyright notice, this list of conditions and the
-  following disclaimer in the documentation and/or other
-  materials provided with the distribution.
-
-* Neither the name of Luke Jennings nor the names of its
-  contributors may be used to endorse or promote products
-  derived from this software without specific prior
-  written permission of Luke Jennings.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED
-WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
-PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
-ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR
-TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
-ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-*/
-#ifndef INC_LIST_TOKENS_H
-#define INC_LIST_TOKENS_H
-
-// Token struct definitions
-typedef struct
-{
-	char username[256];
-	HANDLE token;
-} SavedToken;
-
-typedef struct
-{
-	char username[256];
-	int token_num;
-	BOOL delegation_available;
-	BOOL impersonation_available;
-} unique_user_token;
-
-typedef enum
-{
-	BY_USER,
-	BY_GROUP
-} TOKEN_ORDER;
-
-typedef struct
-{
-	BOOL SE_ASSIGNPRIMARYTOKEN_PRIVILEGE;
-	BOOL SE_CREATE_TOKEN_PRIVILEGE; 
-	BOOL SE_TCB_PRIVILEGE; 
-	BOOL SE_TAKE_OWNERSHIP_PRIVILEGE; 
-	BOOL SE_BACKUP_PRIVILEGE; 
-	BOOL SE_RESTORE_PRIVILEGE; 
-	BOOL SE_DEBUG_PRIVILEGE; 
-	BOOL SE_IMPERSONATE_PRIVILEGE; 
-	BOOL SE_RELABEL_PRIVILEGE; 
-	BOOL SE_LOAD_DRIVER_PRIVILEGE; 
-} TOKEN_PRIVS;
-
-SavedToken *get_token_list(DWORD *num_tokens_enum, TOKEN_PRIVS *token_privs);
-void list_unique_tokens(TOKEN_ORDER);
-void process_user_token(HANDLE, unique_user_token*, DWORD*, TOKEN_ORDER);
-
-#endif
